+++ CHANGELOG.md
@@ -6,6 +6,7 @@ All notable changes to [@bpmn-io/feel-editor](https://github.com/bpmn-io/feel-ed
 
 ___Note:__ Yet to be released changes appear here._
 
+* `FEAT`: enhance FEEL syntax highlighting ([#100](https://github.com/bpmn-io/feel-editor/pull/100))
 * `DEPS`: update to `@camunda/feel-builtins@1.2.0`
 * `DEPS`: update to `min-dom@5.3.0`
 * `DEPS`: update `codemirror*` dependencies